NotFoundException is now thrown in Repository instead of Service

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -9,6 +9,7 @@ import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 import ru.javawebinar.topjava.util.ValidationUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.time.LocalDate;
 import java.util.Collection;
@@ -25,41 +26,54 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> save(meal, 1));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, int userId) {
         log.info("save {}", meal);
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+        }
+        if (repository.get(userId) != null && repository.get(userId).getUserId() != userId) {
+            throw new NotFoundException("this meal is not yours");
         }
         repository.put(meal.getId(), meal);
         return meal;
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, int userId) {
         log.info("delete {}", id);
+        if (repository.get(userId) != null && repository.get(userId).getUserId() != userId) {
+            throw new NotFoundException("this meal is not yours");
+        }
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, int userId) {
         log.info("get {}", id);
+        if (repository.get(userId) != null && repository.get(userId).getUserId() != userId) {
+            throw new NotFoundException("this meal is not yours");
+        }
         return repository.get(id);
     }
 
     @Override
-    public Collection<Meal> getAll() {
+    public Collection<Meal> getAll(int userId) {
+        if (repository.get(userId) != null && repository.get(userId).getUserId() != userId) {
+            throw new NotFoundException("this meal is not yours");
+        }
         return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
                 .sorted((o1, o2) -> o1.getTime().isAfter(o2.getTime()) ? 1 : -1)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public List<Meal> getFilteredByDate(LocalDate startDate, LocalDate endDate) {
-        return this.getAll().stream()
+    public List<Meal> getFilteredByDate(int userId, LocalDate startDate, LocalDate endDate) {
+        return this.getAll(userId).stream()
                 .filter(meal -> DateTimeUtil.isBetween(meal.getDate(), startDate, endDate))
                 .collect(Collectors.toList());
     }
